Extract positive test result notification logic in an injectable service
