+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/notification/NotificationHelper.kt
@@ -17,11 +17,6 @@ import androidx.core.app.NotificationManagerCompat
 import de.rki.coronawarnapp.BuildConfig
 import de.rki.coronawarnapp.CoronaWarnApplication
 import de.rki.coronawarnapp.notification.NotificationConstants.NOTIFICATION_ID
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_ID
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_INITIAL_OFFSET
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_INTERVAL
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_TOTAL_COUNT
-import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.ui.main.MainActivity
 import org.joda.time.Duration
 import org.joda.time.Instant
@@ -91,20 +86,6 @@ object NotificationHelper {
         }
     }
 
-    fun schedulePositiveTestResultReminder(nowUTC: Instant) {
-        if (LocalData.numberOfRemainingPositiveTestResultReminders < 0) {
-            Timber.v("Schedule positive test result notification")
-            LocalData.numberOfRemainingPositiveTestResultReminders = POSITIVE_RESULT_NOTIFICATION_TOTAL_COUNT
-            scheduleRepeatingNotification(
-                nowUTC.plus(POSITIVE_RESULT_NOTIFICATION_INITIAL_OFFSET),
-                POSITIVE_RESULT_NOTIFICATION_INTERVAL,
-                POSITIVE_RESULT_NOTIFICATION_ID
-            )
-        } else {
-            Timber.v("Positive test result notification has already been scheduled")
-        }
-    }
-
     fun cancelFutureNotifications(notificationId: Int) {
         val pendingIntent = createPendingIntentToScheduleNotification(notificationId)
         val manager =
@@ -113,7 +94,7 @@ object NotificationHelper {
         Timber.v("Canceled future notifications with id: %s", notificationId)
     }
 
-    private fun scheduleRepeatingNotification(
+    fun scheduleRepeatingNotification(
         initialTime: Instant,
         interval: Duration,
         notificationId: NotificationId
