+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/settings/SettingsResetViewModel.kt
@@ -5,9 +5,7 @@ import com.squareup.inject.assisted.AssistedInject
 import de.rki.coronawarnapp.exception.ExceptionCategory
 import de.rki.coronawarnapp.exception.reporting.report
 import de.rki.coronawarnapp.nearby.InternalExposureNotificationClient
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_ID
-import de.rki.coronawarnapp.notification.NotificationHelper
-import de.rki.coronawarnapp.storage.LocalData
+import de.rki.coronawarnapp.notification.TestResultNotificationService
 import de.rki.coronawarnapp.ui.SingleLiveEvent
 import de.rki.coronawarnapp.util.DataReset
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
@@ -17,7 +15,8 @@ import de.rki.coronawarnapp.worker.BackgroundWorkScheduler
 
 class SettingsResetViewModel @AssistedInject constructor(
     dispatcherProvider: DispatcherProvider,
-    private val dataReset: DataReset
+    private val dataReset: DataReset,
+    private val testResultNotificationService: TestResultNotificationService
 ) : CWAViewModel(dispatcherProvider = dispatcherProvider) {
 
     val clickEvent: SingleLiveEvent<SettingsEvents> = SingleLiveEvent()
@@ -44,16 +43,11 @@ class SettingsResetViewModel @AssistedInject constructor(
                     ExceptionCategory.EXPOSURENOTIFICATION, TAG, null
                 )
             }
-            resetPositiveTestResultNotification()
+            testResultNotificationService.resetPositiveTestResultNotification()
 
             dataReset.clearAllLocalData()
             clickEvent.postValue(SettingsEvents.GoToOnboarding)
         }
-    }
-
-    private fun resetPositiveTestResultNotification() {
-        NotificationHelper.cancelFutureNotifications(POSITIVE_RESULT_NOTIFICATION_ID)
-        LocalData.numberOfRemainingPositiveTestResultReminders = Int.MIN_VALUE
     }
 
     companion object {
