+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/warnothers/SubmissionResultPositiveOtherWarningViewModel.kt
@@ -6,8 +6,7 @@ import com.squareup.inject.assisted.Assisted
 import com.squareup.inject.assisted.AssistedInject
 import de.rki.coronawarnapp.exception.NoRegistrationTokenSetException
 import de.rki.coronawarnapp.nearby.ENFClient
-import de.rki.coronawarnapp.notification.NotificationConstants.POSITIVE_RESULT_NOTIFICATION_ID
-import de.rki.coronawarnapp.notification.NotificationHelper
+import de.rki.coronawarnapp.notification.TestResultNotificationService
 import de.rki.coronawarnapp.service.submission.SubmissionService
 import de.rki.coronawarnapp.storage.LocalData
 import de.rki.coronawarnapp.storage.interoperability.InteroperabilityRepository
@@ -33,7 +32,8 @@ class SubmissionResultPositiveOtherWarningViewModel @AssistedInject constructor(
     dispatcherProvider: DispatcherProvider,
     private val enfClient: ENFClient,
     private val taskController: TaskController,
-    interoperabilityRepository: InteroperabilityRepository
+    interoperabilityRepository: InteroperabilityRepository,
+    private val testResultNotificationService: TestResultNotificationService
 ) : CWAViewModel(dispatcherProvider = dispatcherProvider) {
 
     private var currentSubmissionRequestId: UUID? = null
@@ -98,7 +98,7 @@ class SubmissionResultPositiveOtherWarningViewModel @AssistedInject constructor(
             submitWithNoDiagnosisKeys()
             routeToScreen.postValue(SubmissionNavigationEvents.NavigateToSubmissionDone)
         }
-        NotificationHelper.cancelFutureNotifications(POSITIVE_RESULT_NOTIFICATION_ID)
+        testResultNotificationService.cancelPositiveTestResultNotification()
     }
 
     private fun submitDiagnosisKeys(keys: List<TemporaryExposureKey>) {
